add player name functionality
Create a function "playerNameScreen()" that creates two inputs
for player name entry, and a button for player names submission.
Once the button is clicked, update gameController player objects with
the respective player name inputs. The input fields and the submission
button will be removed from the DOM after the button click, and
updateScreen() will be invoked to begin gameplay.

Replace initial render with playerNameScreen() to allow players to
input their nicknames.